Session 11 cont: w8a8 triton-3.7 port landed, B300 gate sweep (w8a8 still loses on sm_103), journal recovery, docs sync

+++ profiling/w8a8/w8a8_bmm.py
@@ -69,7 +69,7 @@ def _w8a8_bmm_kernel(
         k_mask = offs_k < K - k * BLOCK_K
         a = tl.load(a_ptrs, mask=m_mask[:, None] & k_mask[None, :], other=0)
         w = tl.load(w_ptrs, mask=k_mask[:, None] & n_mask[None, :], other=0)
-        acc = tl.dot(a, w, acc=acc)  # int8 x int8 -> int32 (IMMA)
+        acc = tl.dot(a, w, acc=acc, out_dtype=tl.int32)  # int8 x int8 -> int32 (IMMA)
         a_ptrs += BLOCK_K * stride_ak
         w_ptrs += BLOCK_K * stride_wk
